Cleanup in classes Sequence, Gene, Protein * Removed m_modified (not used anywhere, was not consistently implemented). * Removed uint16 (the standard lib defines uint, we shouldn't define 	our own type to provide the same functionality). * Added some documentation.
git-svn-id: http://wlab.ccbb.utexas.edu/svn/evoli@126 dadfaeae-b11a-0410-95f2-80a8e7ca9aa5

+++ trunk/src/folder/gene-util.hh
@@ -249,7 +249,7 @@ public:
 		Gene g(p.length()*3);
 		Protein::const_iterator it = p.begin();
 		
-		for (uint16 i=0; it != p.end(); it++, i++) {
+		for (uint i=0; it != p.end(); it++, i++) {
 			int aa = *it;
 			int codon = GeneticCodeUtil::residueToCodonTable[aa+1];
 			g[i] = codon;
